backend/requirements.txt

django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
firebase-admin==6.2.0
requests==2.31.0
razorpay==1.4.1
python-dotenv==1.0.0
